Huge internal update.
November 24 2024: Finished these internal pages: Color Coding. About (for the most part - it is now in a 'working' version available to the public).
Furthermore, I changed the front page logo from one that said "J. Lacabe" to one that says "Human Knowledge", since this is not my personal blog. I left the original "J. Lacabe" image on the About page, for perpetuity. Additionally, every page on the site received an edit to account for the new internal pages, including sidebar additions and changed beginning paragraphs for the pages that reference color coding for the notes.