Updates from Friday
I had an mun conference yesterday, so I was too busy to update the site. Today is the full astronomy day.
December 6 2024: Further work on Calculus. Additionally, every page on the website was updated to incorporate the new Art page, which should be the last major site-wide update until the summer plan (or until I think of something else). The Sitemap was also updated for the first time in a while.
Rules Added: Mathematics, Single Variable Calculus - Rule 109, 110.