Physics + Astronomy
December 18 2024. Adding equations to the Physics equations page, as well as general reformating of the page to make it easier to quickly find equations and definitions. Minor edits to some physics pages naturally came of this as I inspected some rules and noticed typos/minor errors. Furthermore, Astronomy finally received its logo, after a month of deliberations with the mastermind artist of his generation, Shoaib Pasha.