Cryptography & Aerospace
December 20 2024. Work on the Human Knowledge Project can be a slow-moving bureaucracy, but once the pages and all the miscellania done to improve the general look and quality-of-life of users on the website (such as the humorous and/or inspirational quotes under each section), real work is relatively steady. Work on Cryptography had begun. The main pages have been created, and the art for Cryptography has already been created as well. Furthermore, the "Aerospace Engineering" logo was created in advance of the page actually being created, making today extremely productive in terms of Project Art.