DES, Calculus
December 31 2024: Work on Cryptography, developing the DES section significantly, but also some on Calculus (in my own notebook, to be transferred to the site soon). By the end of 2025, I hope for all of Calculus, Multivariable Calculus, and Linear Algebra to be finished on the Math side, all of Electromagnetism (and some entry into Quantum Physics) on the Physics side, and for all of the currently established Cryptography sections (Symmetric Cryptography through Post-Quantum) to be finished as well. A tall order, but there is a considerable amount of time between then and now.
Rules Added: Cryptography, Symmetric Cryptography - Rule 43, 44, 45, 46.