Rule fixes + Physics
March 1 2025: Upon reviewing notes, I found several parts difficult to understand. In the process of learning the material, I tend to add content sequentially, which may pose a challenge to my future self if I find that the content is difficult to understand. As such,
Rules Added: Physics, Electromagnetism - Rule 182, 183.