Physics work.
March 9 2025: Physics work, to a degree that has not been seen in some time and what must be replicated everyday in the future - akin to what was accomplished during late December/early January for Cryptography. I have come to a level of confusion on a particular subject (that of the internal electric field of a nonconducting shell), only exacerbated by the supposed "educational tool" of the internet: multitudes of contradicting answers and conflicting explanations, from artificial intelligence, to the incomplete and vague explanation in the textbooks, to those given on Physics forums by the woefully incompetent and neglectful of the fact that net zero electric flux does not mean net zero electric field, nor does it necessarily imply it. Thus only goes to show the inherant value of the Human Knowledge Project in presenting an unadulterably true and absolute description and elaboration on scientific/physical concepts, backed up purely using the already known principles and equations. Cutting through the noise to present a simple, intuitive understanding - this will realize the internet for its highest and greatest possible purpose, a potential, as shown, that is currently unfulfilled. I will have to seek help from an expert on the topic, of whom the closest to which will be my physics teacher.
Rules Added: Physics, Electromagnetism - Rule 193, 194.