Electric Fields completed.
March 10 2025: A total revision of Physics Rule 193, which led to the creation of a new Rule dealing with Asymmetric Charge Distribution, as to cover a topic that was becoming too large to be all described within Rule 193 itself. That, alongside the addition of a new subsection to deal with the new electric fields created using the Charge-Flux Law that were impossible to derive before.
Rules Added: Physics, Electromagnetism - Rule 195.