Capacitance Work
April 1 2025: Very productive day for Physics - With every addition to the new sections (Capacitance), I find I have to go back and revise a previous section in some way, while still keeping in mind the level of knowledge that had been reached by that section, and thus referring to future knowledge sparingly and with pointing references if needed. In any case, the dragon that is the dielectric Rules has been conquered, and the sun has almost totally set on Capacitance as a topic.
Rules Added: Physics, Electromagnetism - Rule 215, 216, 217.