Electromagnetism Equations
April 8 2025: Considerable work in the Physics Equations page, incorporating most of Section XII. I have also reached a very exciting stage in circuitry and Electromagnetism as a whole - whereas before I found myself afflicted with a sense of confusion with several concepts, I now (after working very hard to understand them) find the rest of the remaining material much more tedious than confusing & laborious with regard to incorporating it into the website, which makes the creation of the notes not particularly a matter of huge concentration and brainpower, but rather one of simple repetition akin to maintenance work - there are very few new questions to be asked at this specific point. It is almost as if Circuitry has become an exhausted subject.