Electromagnetism - Series & Parallel
April 17 2025 - The treatise on Series and Parallel sets of Resistors and Capacitors has been completed. I included both Full Analysis and Truncated version, for ease of reference depending on whether one is studying the concept as a whole or just needs a refresher on the gist of the concept.
Rules Added: Physics, Electromagnetism - Rule 245, 246, 247, 248.

April 16 2025 - Thorough revisals in Electromagnetism, such as the "sludge" analogy of Rule 238. Additionally, a minor fix was made to the About page.