Physics, Math Logo, Spoilers
June 14th, 2025: Extensive Physics Revisals. Additionally, after over five months since the last logo addition, the beautiful Mathematics logo has finally been created and added to the site. See it here: https://humanknowledge.neocities.org/_other/logos/Mathematics.png. This logo required outside attribution as a result of its incorporation of a Vecteezy jpg. Furthermore, on the programming side of things, I further kludged the already weirdly formatted css to incorporate Spoilers, which I came up with a need for as a result of the answers for example problems, a small number of which exist in the Rules (the "Dimensional Analysis Comprehension Problem" of Classical Mechanics for example). In classic CSS fashion, this change took forever to be reflect onto the pages using it.