Modular Arithmetic work.
July 9th, 2025: Rules 8 and 9 in Symmetric Cryptography (relating to Modular Arithmetic) were completely revised, and the two were particularly troublesome considering the inadequacy of the previous writings for them. Additionally, I have begun preparing for newer Physics subcategories, such as Vibrations and Waves. There is no logo yet, but I have already prepared a quote for it on the main Physics page.