Math Symbols + Cryptography
July 25th, 2025: Further work in Cryptography and the math symbols page. I completely overhalled the math symbols page, which was long overdue for linkings allowing each symbol to be referenced. As for the incomplete portion of the Constants section, I intend to finish that by the end of the summer. In working on Cryptography, I realized that my planned Subcategories for future work did not include Elliptic Curve Cryptography, and so I added the requisite pages and made edits as necessary (such as to the equations page, and to the bottom links of each subcategory page). In doing so, I also created the Cult of Jerry™ (see the source code of the Cryptography page).

July 24th, 2025: Further Cryptography work. Additionally, the Single Variable Calculus logo was added.