Site-wide error fix
July 26th, 2025: Cryptography work, with the creating of a new section in Stream Ciphers. Additionally, I performed another site-wide minor sidebar change after I saw the website from a friend's computer and found that there were some design oversights that flew over the radar because it looked fine on _my computer in particular_. (*sigh*) I additionally noted that images with links attached to them (e.g. every image using the summaries) had the dotted blue link boxes that appear when hovering malformed, with the boxes only showing at the bottom of the image with the same height as a textlink (again: worked fine on my machine). Only a simple fix to the css was required to rectify this disaster.