Internal work
July 28th, 2025: More internal work. In order to fix the issue of image with links having the dotted blue boxes rendered differently on different browsers (as a result of the css working differently on Firefox than Chrome, as well as others (shout out to browserling.com for helping determine which browsers looked different!)), I spend some time fiddling with the css to make linked images look the same cross-browser - this was accomplished with some effort. I also made some other changes to linking in general, with a new fade-in time and a darker shade of green when hovering. This required *another* sitewide update to fix all image with links, which, due to the sidelinks, effects the entire website. After this debacle, I discovered that every directory on the website (/Physics/, /information/, etc.) was publically-facing on the main hkproject.org website - in order to fix this, I just created an index.html for each directory, which I just set to the not found page as a default. I also made several edits to the not found page itself.