Block Cipher work
August 1, 2025: I was so busy yesterday that I forgot that it was the 1 year anniversary of the project. In celebration, I spent the day working on Cryptography. Perhaps I will plan something more exciting next year. Today, I updated the sitemap. Further work on Cryptography, of course, densely concentrated in Block Ciphers.